refactor(web_fetch): extract options and URL rules
